deployment/azure-python-deploy-to-all-existing-storage/deployToAllExistingStorageAccounts/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
azure-cli==2.37.0
azure-functions==1.7.2
azure-storage-blob==12.12.0
azure-identity==1.10.0 ## azure-cli satisfies this requirement
azure-mgmt-resource==21.1.0b1 ## azure-cli satisfies this requirement
azure-mgmt-subscription==3.0.0
azure-mgmt-storage==20.0.0
azure-keyvault-secrets==4.4.0
urllib3==1.26.9
argparse==1.4.0
uuid7==0.1.0
munch==2.5.0